==> 02_Get_NBA_stats.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import time

from selenium import webdriver
from selenium.webdriver.chrome.options import Options 

chrome_path = '/Users/MayankPhanse/Desktop/chromedriver'

options = Options() 

# set the window size
options.add_argument('--window-size=500,300')

# initialize the driver
driver = webdriver.Chrome(chrome_path, 
                          options=options)

driver.set_window_size(1400,1000)

driver.get('https://stats.nba.com/teams/boxscores-traditional/?Season=2019-20&SeasonType=Regular%20Season')

## Get First Half Stats

def get_first_half_stats():
    driver.get('https://stats.nba.com/teams/boxscores-traditional/?Season=2019-20&SeasonType=Regular%20Season')
    #clicks advanced filters
    driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/div[1]/div[4]/a').click()
    time.sleep(1)
    #clicks first half
    driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/div[2]/div[1]/div[2]/div[5]/div/label/select/option[2]').click()
    time.sleep(1)
    #clicks run it
    driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/div[2]/div[2]/stats-run-it/a').click()
    time.sleep(1)
    #clicks all
    driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[1]/div/div/select/option[1]').click()
    
    page_source = driver.page_source
    soup = BeautifulSoup(page_source)
    tables = soup.findAll('table')
    table = tables[0].prettify()
    first_half = pd.read_html(table)[0]

#     first_half.to_csv('first_half.csv')
    return first_half

get_first_half_stats().to_csv('first_half_test.csv')

## Get Second Half Stats

def get_second_half_stats():
    driver.get('https://stats.nba.com/teams/boxscores-traditional/?Season=2019-20&SeasonType=Regular%20Season')
    #clicks advanced filters
    driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/div[1]/div[4]/a').click()
    time.sleep(1)
    #clicks second half
    driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/div[2]/div[1]/div[2]/div[5]/div/label/select/option[3]').click()
    time.sleep(1)
    #clicks run it
    driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/div[2]/div[2]/stats-run-it/a').click()
    time.sleep(1)
    #clicks all
    driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[1]/div/div/select/option[1]').click()
    
    page_source = driver.page_source
    soup = BeautifulSoup(page_source)
    tables = soup.findAll('table')
    table = tables[0].prettify()
    second_half = pd.read_html(table)[0]

    second_half.to_csv('second_half.csv')

get_second_half_stats()

def get_game_stats():
    driver.set_window_size(1400,1000)
    
    driver.get('https://stats.nba.com/teams/boxscores-traditional/?Season=2019-20&SeasonType=Regular%20Season')
    
    #clicks all
    driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[1]/div/div/select/option[1]').click()
    
    page_source = driver.page_source
    soup = BeautifulSoup(page_source)
    tables = soup.findAll('table')
    table = tables[0].prettify()
    entire_game = pd.read_html(table)[0]

    entire_game.to_csv('nba_games.csv')

get_game_stats()

**If the paths don't work, just copy and paste the exact links over the previous links and then run**


page_source = driver.page_source
soup = BeautifulSoup(page_source)
tables = soup.findAll('table')
table = tables[0].prettify()
entire_game = pd.read_html(table)[0]

df.to_csv('nba_games.csv')
