==> 03_Clean_Data_For_Usage.ipynb <==
# Data Cleaning for Regular Stats

import pandas as pd

df = pd.read_csv('first_half.csv')

### Read Data

df.head()

df.columns

df.head()

df['Team1'] = df['Match\xa0Up'].str.split().apply(lambda x: x[0])
df['Team2'] = df['Match\xa0Up'].str.split().apply(lambda x: x[2])

(df['Team'] == df['Team1']).all()

df.head()

df.columns

### Get Full Game Stats to set Win Margins

df_game = pd.read_csv('nba_games.csv')

df_game.head()

df_game.sort_values(by = ['Game\xa0Date','Match\xa0Up'],inplace = True)

df_game.reset_index(inplace = True,drop = True)

df_game.head()

df.sort_values(by = ['Game\xa0Date','Match\xa0Up'],inplace = True)
df.reset_index(inplace = True, drop = True)

df.head()

df['End Game +/-'] = df_game['+/-']

df.head()

### Drop Unnecessary or Unused Information

df.drop(['Unnamed: 0','Season','Match\xa0Up','Game\xa0Date','W/L','MIN'], axis = 1,inplace = True)

df.columns

df.head()

df

### Set Target for Spreads

df['win_target'] = df['End Game +/-'] > 0
df['3_win_target'] = df['End Game +/-'] > 3
df['5_win_target'] = df['End Game +/-'] > 5
df['7_win_target'] = df['End Game +/-'] > 7
df['10_win_target'] = df['End Game +/-'] > 10

df.head()

df.drop(['End Game +/-'],axis = 1,inplace=True)

df.head()

**Save Dataframe**

df.to_csv('halftime_df.csv')

### Create Cleaning Function Following Same Steps

def clean_data(csv_file):
    df = pd.read_csv(csv_file)
    
    #set teams
    df['Team1'] = df['Match\xa0Up'].str.split().apply(lambda x: x[0])
    df['Team2'] = df['Match\xa0Up'].str.split().apply(lambda x: x[2])
    
    #delete unused columns
    df.drop(['Unnamed: 0','Season','Match\xa0Up','Game\xa0Date','W/L','MIN'], axis = 1,inplace = True)
    
    #set targets
    df['win_target'] = df['+/-'] > 0
    df['3_win_target'] = df['+/-'] > 3
    df['5_win_target'] = df['+/-'] > 5
    df['7_win_target'] = df['+/-'] > 7
    df['10_win_target'] = df['+/-'] > 10
    
    #drop last unused column
    df.drop(['+/-'],axis = 1,inplace=True)
    
    return df

clean_df = clean_data('nba_games.csv')

clean_df

clean_df.to_csv('clean_nba_games.csv')
