==> 05_Real-Time Test.ipynb <==
import pandas as pd
import numpy as np
import util
import joblib

def get_ht_stats():
    
    team_dict = {'ATL' : 0,
             'BKN' : 1,
             'BOS' : 2,
             'CHA' : 3,
             'CHI' : 4,
             'CLE' : 5,
             'DAL' : 6,
             'DEN' : 7,
             'DET' : 8,
             'GSW' : 9,
             'HOU' : 10,
             'IND' : 11,
             'LAC' : 12,
             'LAL' : 13,
             'MEM' : 14,
             'MIA' : 15,
             'MIL' : 16,
             'MIN' : 17,
             'NOP' : 18,
             'NYK' : 19,
             'OKC' : 20,
             'ORL' : 21,
             'PHI' : 22,
             'PHX' : 23,
             'POR' : 24,
             'SAC' : 25,
             'SAS' : 26,
             'TOR' : 27,
             'UTA' : 28,
             'WAS' : 29}
    
    df = pd.DataFrame(columns = ['Team','PTS','FGM','FGA','FG%','3PM','3PA','3P%','FTM','FTA','FT%','OREB','DREB','REB','AST','TOV','STL','BLK','PF','+/-'],index = [0])
    df['Team'] = input('Enter Team: ')
    df['Team'] = df['Team'].map(team_dict)
    df['PTS'] = int(input('Enter PTS: '))
    df['FGM'] = int(input('Enter FGM: '))
    df['FGA'] = int(input('Enter FGA: '))
    df['FG%'] = df['FGM']/df['FGA']
    df['3PM']= int(input('Enter 3PM: '))
    df['3PA'] = int(input('Enter 3PA: '))
    df['3P%'] = df['3PM']/df['3PA']
    df['FTM'] = int(input('Enter FTM: '))
    df['FTA'] = int(input('Enter FTA: '))
    df['FT%'] = df['FTM']/df['FTA']
    df['OREB'] = int(input('Enter OREB: '))
    df['DREB'] = int(input('Enter DREB: '))
    df['REB'] = df['OREB']+df['DREB']
    df['AST'] = int(input('Enter AST: '))
    df['TOV'] = int(input('Enter TOV: '))
    df['STL']= int(input('Enter STL: '))
    df['BLK'] = int(input('Enter BLK: '))
    df['PF'] = int(input('Enter PF: '))
    df['+/-'] = int(input('Enter +/-: '))
    
    return(df)


model_w = joblib.load('model_w.sav')
model_3 = joblib.load('model_3.sav')
model_5 = joblib.load('model_5.sav')
model_7 = joblib.load('model_7.sav')
model_10 = joblib.load('model_10.sav')

def get_bet():
    df = get_ht_stats()
    
    df_bet = pd.DataFrame(columns = ['Win','Win by 3','Win by 5','Win by 7','Win by 10'], index = [0,1])
    
    df_bet['Win'][0] = model_w.predict(df)
    df_bet['Win by 3'][0] = model_3.predict(df)
    df_bet['Win by 5'][0] = model_5.predict(df)
    df_bet['Win by 7'][0] = model_7.predict(df)
    df_bet['Win by 10'][0] = model_10.predict(df)
    
    df_bet['Win'][1] =  np.around(model_w.predict_proba(df)[:,1]*100,decimals = 1)
    df_bet['Win by 3'][1] = np.around(model_3.predict_proba(df)[:,1]*100,decimals = 1)
    df_bet['Win by 5'][1] = np.around(model_5.predict_proba(df)[:,1]*100,decimals = 1)
    df_bet['Win by 7'][1] = np.around(model_7.predict_proba(df)[:,1]*100,decimals = 1)
    df_bet['Win by 10'][1] = np.around(model_10.predict_proba(df)[:,1]*100,decimals = 1)
    
    return(df_bet)

get_bet()

get_bet()
